+++ src/main/java/com/ManifestTeswTancis/Entity/ExportHouseBlEntity.java
@@ -84,7 +84,7 @@ public class ExportHouseBlEntity implements Serializable {
     private String blDescription;
 
     @Column(name="BL_PACKAGE")
-    private Double blPackage;
+    private Integer blPackage;
 
     @Column(name="PACKAGE_UNIT")
     private String packageUnit;
