+++ src/main/java/com/ManifestTeswTancis/Entity/MasterBlGoodItemsEntity.java
@@ -49,7 +49,7 @@ public class MasterBlGoodItemsEntity implements Serializable {
     private String packingType;
 
     @Column(name = "GOOD_PACKAGE_QUANTITY")
-    private Double packageQuantity;
+    private Integer packageQuantity;
 
     @Column(name = "GOOD_PACKAGE_TYPE")
     private String packageType;
